Fixed OX-3905 in OpenX 2.7.31-beta.
Manually changing account while in statistics screens will now re-direct user to an appropriate page in the statistics system if they do not have permission to view the existing page after switching. Manually entering a URL that the user/account has no permission to view also now correctly shows an error message that permission is denied, and allows the user to log in as a different user.