zh_CN/*.lang.php re-worked
The legacy simplified Chinese i18n files from OpenX 2.8 were very much buggy.
These files are translated based on the latest version of en/*.lang.php, and saved with utf-8 encoding.
This is still an imcomplete version. But it is sufficient for working on most daily routines as the 'Default Manager' role.