Update advertiser-index.php
The $banner variable in the foreach reference loop must be unset to avoid an active advertiser appearing inactive when the $banner variable is used for building the advertiser tree. 

See the warning on http://php.net/manual/en/control-structures.foreach.php :
"Reference of a $value and the last array element remain even after the foreach loop. It is recommended to destroy it by unset()."