Process extra parameters when generating tags
Because the generateTags() call was expecting a struct or array as the third "params" argument, the use of getScalarValues resulted in these parameters being completely ignored. This change uses getNotRequiredScalarValue() to correctly parse the incoming struct or array parameters and pass them on.

This allows the XML call to include $aParams values like array('comments' => 0) to disable comments in the resulting invocation tag, etc.