Update async.js correcting empty space after iframe
By the asycnronous delivery occuring an empty space after the iframe as iframe is an inline HTML-Element. The solution could be in 2 ways: adding display:block (than you need to add some more rules) or adding vertical-align:bottom CSS rule, which removes this empty space. As Google DFP is using the vertical-align rule. I added this to the async.js file.